Delete python/cuda/memory_pool.{h,cc} and move content to CUDA module binding
